requirements.txt
ncatbot==4.4.1.post1
httpx>=0.28.1
PyYAML>=6.0.2
